Testing: Split export tests into multiple files
